fix nnum bug in fix_illum_headers with multiple raw files
This occured for GSE43691 where raw data is split exactly between two files so in
a single file there is 96 numeric columns (p-values and signal) and 96 eset samples.
This resulted in fix_illum_headers thinking there are only enough numeric columns for signal
for each sample and so all were renamed as signal columns.